+++ packages/selector-specificity/test/test.mjs
@@ -1,5 +1,5 @@
 import assert from 'assert';
-import selectorSpecificity from '@csstools/selector-specificity';
+import { selectorSpecificity } from '@csstools/selector-specificity';
 import parser from 'postcss-selector-parser';
 
 function calculate(selector) {